refactor(debug-trace-server): give the middleware stack its own module
