+++ bin/debug-trace-server/src/main.rs
@@ -62,10 +62,6 @@ use stateless_core::{
 use stateless_db::ContractCache;
 use tokio::task;
 use tokio_util::sync::CancellationToken;
-use tower::{
-    ServiceBuilder,
-    layer::util::{Identity, Stack},
-};
 use tracing::{debug, error, info, instrument, warn};
 
 mod block_data_cache;
@@ -73,6 +69,7 @@ mod chain_sync;
 mod compression;
 mod data_provider;
 mod metrics;
+mod middleware;
 mod response_cache;
 mod response_size;
 mod rpc_service;
@@ -404,31 +401,6 @@ fn validate_args(args: &Args) -> Result<()> {
     Ok(())
 }
 
-/// The middleware stack [`http_middleware`] composes, innermost layer first.
-type HttpMiddleware = ServiceBuilder<
-    Stack<
-        timing::TimingHeaderLayer,
-        Stack<
-            response_size::ResponseSizeLayer,
-            Stack<compression::ResponseCompressionLayer, Identity>,
-        >,
-    >,
->;
-
-/// Composes the HTTP middleware stack — the one place its order is defined (the unit
-/// tests in [`compression`] run requests through this exact stack).
-///
-/// Compression must stay outermost: `ResponseSizeLayer` reads the body's exact
-/// `size_hint` (gone once the body is a compressed stream), and the timing layer's
-/// future resolves before the body streams — so this order keeps `x-response-size` at
-/// the uncompressed payload size and `x-execution-time-ns` free of compression CPU.
-pub(crate) fn http_middleware(compression_enabled: bool) -> HttpMiddleware {
-    ServiceBuilder::new()
-        .layer(compression::layer(compression_enabled))
-        .layer(response_size::ResponseSizeLayer)
-        .layer(timing::TimingHeaderLayer)
-}
-
 #[tokio::main]
 async fn main() -> Result<()> {
     let args = Args::parse();
@@ -658,7 +630,7 @@ async fn main() -> Result<()> {
     let config = ServerConfig::builder().max_response_body_size(u32::MAX).build();
     let server = Server::builder()
         .set_config(config)
-        .set_http_middleware(http_middleware(!args.response_compression_disabled))
+        .set_http_middleware(middleware::http_middleware(!args.response_compression_disabled))
         .build(&args.addr)
         .await?;
     let addr = server.local_addr()?;
